refactor: split handler.go into multiple focused handler files
- Extract message event handling to message_handler.go (512 lines)
- Extract interaction handling to interaction_handler.go (247 lines)
- Extract slash command handling to slash_command_handler.go (65 lines)
- Extract message posting utilities to message_utils.go (97 lines)
- Reduce handler.go from 1300+ lines to 211 lines (core structure only)

This refactoring improves code organization by separating concerns:
- Each handler type now has its own file
- Core Handler struct and configuration remain in handler.go
- File/image processing is kept with message handling where it belongs
- All tests pass without modification
